apps/bg-remover/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
rembg[cpu]>=2.0.55
pillow>=10.2.0
